Minor refactoring

+++ tables.c
@@ -2604,7 +2604,7 @@ fixed_t finetangent(angle_t angle) {
 fixed_t finesine(angle_t angle) {
 	int o, res;
 	unsigned q;
-	static int16_t offset[4] = { 0, 16, -16, 32 };
+	static int8_t offset[4] = { 0, 16, -16, 32 };
 	const unsigned short *s;
 
 	angle &= FINEMASK;
